% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{benchmark_portfolio}
\alias{benchmark_portfolio}
\title{benchmark_portfolio}
\usage{
benchmark_portfolio(df, wght, rebalance, benchmark_column, from, to)
}
\arguments{
\item{weights, }{change, key parameters.}
}
\description{
benchmark_portfolio
}
