% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{calcPortPerformance}
\alias{calcPortPerformance}
\title{Calculates portfolio return/risk given mean returns of assets and a covariance matrix}
\usage{
calcPortPerformance(weights, mean_ret, cov_matrix)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
Calculates portfolio return/risk given mean returns of assets and a covariance matrix
}
