% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{calculate_optimal_weights}
\alias{calculate_optimal_weights}
\title{calculate_optimal_weights}
\usage{
calculate_optimal_weights(instruments, bt_df, from, to, scale = 252)
}
\arguments{
\item{instruments, }{bt_df, from, to, scale parameters.}
}
\description{
calculate_optimal_weights
}
