% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{calculate_optmised_returns}
\alias{calculate_optmised_returns}
\title{Build optimal portfolios returns}
\usage{
calculate_optmised_returns(instruments, from, to, opt_w, port_ret)
}
\arguments{
\item{instruments, }{from, to, opt_w, port_ret parameters.}
}
\description{
Build optimal portfolios returns
}
