% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{calculate_risk_metrics}
\alias{calculate_risk_metrics}
\title{Calculates portfolio performance measures}
\usage{
calculate_risk_metrics(returns, risk_free, from, to, scale = 252)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
Calculates portfolio performance measures
}
