% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{findEfficientFrontier.Return}
\alias{findEfficientFrontier.Return}
\title{Finds weights of assets on the efficient frontier by target return}
\usage{
findEfficientFrontier.Return(returns, target_ret, short = FALSE, scale = 252)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
Finds weights of assets on the efficient frontier by target return
}
