% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{findEfficientFrontier.ReturnALT}
\alias{findEfficientFrontier.ReturnALT}
\title{By target return ALTERNATIVE}
\usage{
findEfficientFrontier.ReturnALT(mean_ret, cov_matrix, target_ret)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
By target return ALTERNATIVE
}
