% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{findEfficientFrontier.Risk}
\alias{findEfficientFrontier.Risk}
\title{By target risk}
\usage{
findEfficientFrontier.Risk(mean_ret, cov_matrix, target_risk)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
By target risk
}
