% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasources.R
\name{fred_to_daily_rates}
\alias{fred_to_daily_rates}
\title{Convert FRED rates to daily interest rates}
\usage{
fred_to_daily_rates(a)
}
\arguments{
\item{a}{FRED interest rate.}
}
\description{
Convert FRED rates to daily interest rates
}
