% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasources.R
\name{load_fred_data}
\alias{load_fred_data}
\title{Load economic datasets from FRED website}
\usage{
load_fred_data(symbols, start_date, end_date)
}
\arguments{
\item{symbols, }{start_date, end_date Fred dataset and period.}
}
\description{
Load economic datasets from FRED website
}
