% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasources.R
\name{load_instruments}
\alias{load_instruments}
\title{Load default list of instruments}
\usage{
load_instruments(cache_folder)
}
\arguments{
\item{cache_folder}{folder for saving the downloaded data.}
}
\description{
Load default list of instruments
}
