% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasources.R
\name{load_tickers}
\alias{load_tickers}
\title{Load tickers from yahoo finance}
\usage{
load_tickers(symbols, start_date, end_date, cache_folder)
}
\arguments{
\item{symbols, }{start_date, end_date, cache_folder tiker symbols and period.}
}
\description{
Load tickers from yahoo finance
}
