% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{MOD_SHINY_FINANCE_NAMESPACE}
\alias{MOD_SHINY_FINANCE_NAMESPACE}
\title{PACKAGE CONSTANTS}
\format{
An object of class \code{character} of length 1.
}
\usage{
MOD_SHINY_FINANCE_NAMESPACE
}
\description{
PACKAGE CONSTANTS
}
\keyword{datasets}
