% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{portfolio_returns}
\alias{portfolio_returns}
\title{Calculates portfolio returns}
\usage{
portfolio_returns(df, from, to, weights, rebalance, geometric = TRUE)
}
\arguments{
\item{df, }{from, to, weights, rebalance, geometric parameters.}
}
\description{
Calculates portfolio returns
}
