% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{rebalance}
\alias{rebalance}
\title{Rebalance weights vector given a change so the sum is 1}
\usage{
rebalance(weights, change, key)
}
\arguments{
\item{weights, }{change, key parameters.}
}
\description{
Rebalance weights vector given a change so the sum is 1
}
