% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolio.R
\name{simPortfolios}
\alias{simPortfolios}
\title{Simulates portfolio risk/return given mean returns of assets and a covariance matrix}
\usage{
simPortfolios(mean_ret, cov_matrix, nsim = 10000)
}
\arguments{
\item{returns, }{risk_free, from, to, scale.}
}
\description{
Simulates portfolio risk/return given mean returns of assets and a covariance matrix
}
